Fix setting default entry on ppc64le when using OPAL
The GRUB menuentry command allows to separate the arguments for options
using either a '=' or a ' '. The latter is the convention used when the
menu entries are defined in the GRUB config file, but this is currently
not supported by Petitboot.

So as a workaround define the menu entries using '--id=${bls}' instead.

Resolves: rhbz#1721815

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>