Update gating.yaml to match the latest BaseOS CI test types
The baseos-ci.brew-build.tier1-other-archs.functional was used to test for
all non-x86_64 arches. But now it was split in three different test groups:

 - baseos-ci.brew-build.tier1-ppc64le-of.functional
 - baseos-ci.brew-build.tier1-ppc64le-opal.functional
 - baseos-ci.brew-build.tier1-aarch64-efi.functional

The first two are for tests that run on ppc64le using OF and OPAL firmware
interfaces respectively, and the last one for test that are run on aarch64.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>