Build and sign powerpc-ieee1275 images
In the ppc64le Secure Boot chain, the boot firmware verifies the GRUB core
image using an appended signature, which is a signature of the whole image
wrapped in a PKCS#7 signedData message and appended to the signed file.

This patch builds an image with the needed modules and sign it.

A signed image is installed in /usr/lib/grub/powerpc-ieee1275/core.elf
and can be used to update the PReP partition. This is similar to how
signed PE32+ binaries are shipped for a UEFI Secure Boot configuration.

The same appended signatures format is used by GRUB to verify the booted
Linux kernels. A x509 certificate with the public key is embedded in the
GRUB core image to be used for this verification.

Multiple certificates can be embedded in GRUB for Linux kernel signature
validation, so add both the "redhatsecureboot303" currently used and the
"redhatsecureboot601" that likely will be used in future for consistency.

Related: rhbz#1899864

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>