Go back to installing unicode.pf2
lorax has its own code for building EFI images, and it needs the
font file to do that successfully, so let's make sure it's there
for lorax to find. This doesn't revert the embedding change,
it just reverts the part where we don't bother to install the
font to /boot/grub2/fonts any more.

Signed-off-by: Adam Williamson <awilliam@redhat.com>