Ensure all scriptlets always exit 0, per guidelines
Fedora package scriptlets are required to always exit 0. grub2's
do not currently ensure this, and especially the new %posttrans
scriptlet for grub2-efi actually exits non-zero during ostree
builds, causing them to fail:
https://bodhi.fedoraproject.org/updates/FEDORA-2025-5a25808a78#comment-3947404
This should fix that, and also make sure we never get the same
problem with the other scriptlets in the package.

Signed-off-by: Adam Williamson <awilliam@redhat.com>