20-grub.install: Skip BLS removal when entry type is type2
UKI(package is kernel-uki-virt) is a single, bootable file that bundles everything needed to start a Linux system.
It contains its own bootable stub and bypasses GRUB2 completely.
The kernel-core and kernel-uki-virt can coexist in one machine. And both of them call kernel-install remove <kversion> upon package
removal and this leads to the complete removal of both the traditional kernel & its artifacts(initramfs, BLS entry file,...).
For example, if the customer remove kernel-uki-virt, currently it also removes BLS entry which causes the regular kernel fails to boot up.
In https://github.com/systemd/systemd/pull/37897 it added --entry-type=type1|type2 option to kernel-install. type1 stands for normal kernel,
type2 stands for uki. When kernel-install is invoked with --entry-type=type2 which is for UKI, we should not remove the BLS entry.

Signed-off-by: Yuxin Sun <yuxisun@redhat.com>