55-set-boot-entry.install: fix initrd check
Fix where to look for initrd file, as /boot may not be a separate
partition, in which case adding /boot unconditionally leads to
a duplicate "/boot/boot/...".

Fix initrd line parsing to take only first word - it can contain
extra parameters, e.g. "$tuned_initrd", which are not part
of initrd filename.

Resolves: #2386118
Reported-by: Mikhail Gavrilov <mikhail.v.gavrilov@gmail.com>
Tested-by: Mikhail Gavrilov <mikhail.v.gavrilov@gmail.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>